task 1 init impl

+++ src/main/kotlin/day_17/Common.kt
@@ -1,6 +1,6 @@
 package day_17
 
-data class Computer(val A: Long, val B: Long, val C: Long, val program: List<Byte>)
+data class Computer(val A: Int, val B: Int, val C: Int, val program: List<Int>)
 
 fun transform(input: String): Computer {
     val blocks = input.split("\n\n")
@@ -8,11 +8,11 @@ fun transform(input: String): Computer {
     val A =  register(regs[0])
     val B =  register(regs[1])
     val C =  register(regs[2])
-    val program = blocks[1].substringAfterLast(" ").split(",").map { it.toByte() }
+    val program = blocks[1].substringAfterLast(" ").split(",").map { it.toInt() }
     return Computer(A, B, C, program)
 }
 
-fun register(data: String) = data.substringAfterLast(" ").toLong()
+fun register(data: String) = data.substringAfterLast(" ").toInt()
